Database connnection using ADO.net and displaying in view done

+++ POCOLib/Product.cs
@@ -19,7 +19,7 @@ namespace ECommerseEntities
 
         public string Description { get; set; }
 
-        public int UnitPrice { get; set;}
+        public double UnitPrice { get; set;}
 
         public int Quantity { get; set;}
         public string ImgUrl { get; set;}
